Add Little endian and big endian helpers
